@@ -8,10 +8,10 @@
 class MergeSort : virtual  public Sort
 {
 private:
-    void merge(int const left, int const mid, int const right);
-    void _mergeSort(int const begin, int const end);
+    void merge(int const t_left, int const t_mid, int const t_right);
+    void mergeSort(int const t_begin, int const t_end);
 public:
-    MergeSort(std::vector<int>& arr);
+    MergeSort(std::vector<int>& t_arr);
 
     void sort() override;
 };
