@@ -8,10 +8,10 @@
 class QuickSort : virtual public Sort
 {
 private:
-    int partition(std::vector<int>& arr, int low, int high);
-    void quickSort(std::vector<int>& arr, int low, int high);
+    int partition(int t_low, int t_high);
+    void quickSort(int t_low, int t_high);
 public:
-    QuickSort(std::vector<int>& arr);
+    QuickSort(std::vector<int>& t_arr);
 
     void sort() override;
 };
