@@ -1,10 +1,10 @@
 #include "sort/CombSort.h"
 
-CombSort::CombSort(std::vector<int>& arr) : Sort(arr) {}
+CombSort::CombSort(std::vector<int>& t_arr) : Sort(t_arr) {}
 
-int CombSort::getNextGap(int gap)
+int CombSort::getNextGap(int t_gap)
 {
-    gap = (gap*10)/13;
+    int gap = (t_gap * 10) / 13;
     if (gap < 1)
         return 1;
     return gap;
@@ -40,7 +40,7 @@ void CombSort::sort()
         swapped = false;
  
         // Compare all elements with current gap
-        for (int i = 0; i < elems.size() - gap; i++)
+        for (size_t i = 0; i < elems.size() - gap; i++)
         {
             if (elems[i] > elems[i + gap])
             {
